(fix) move all connector tests to connectors folder
