+++ test/connectors/quickswap/quickswap.route.test.ts
@@ -1,9 +1,9 @@
 import request from 'supertest';
-import { patch, unpatch } from '../../../services/patch';
-import { gatewayApp } from '../../../../src/app';
-import { Polygon } from '../../../../src/chains/polygon/polygon';
-import { Quickswap } from '../../../../src/connectors/quickswap/quickswap';
-import { patchEVMNonceManager } from '../../../evm.nonce.mock';
+import { patch, unpatch } from '../../services/patch';
+import { gatewayApp } from '../../../src/app';
+import { Polygon } from '../../../src/chains/polygon/polygon';
+import { Quickswap } from '../../../src/connectors/quickswap/quickswap';
+import { patchEVMNonceManager } from '../../evm.nonce.mock';
 let polygon: Polygon;
 let quickswap: Quickswap;
